moar work on float'

+++ core/src/argon/codegen/cppgen/CppGenFltPt.scala
@@ -8,7 +8,7 @@ trait CppGenFltPt extends CppCodegen {
   import IR._
 
   override protected def remap(tp: Type[_]): String = tp match {
-    case FloatType()  => "double"
+    case FloatType()  => "float"
     case DoubleType() => "double"
     case _ => super.remap(tp)
   }
